+++ DAL/Entities/Question.cs
@@ -12,7 +12,7 @@ namespace DAL.Entities
         public int Id { get; set; }
         public string Description { get; set; }
         public bool OneCorrectType { get; set; }
-        public float Points { get; set; }
+        public int Points { get; set; }
         [Required]
         public virtual Topic Topic { get; set; }
         [Required]
